Update calc_scalar in all_functions.R
A scaling factor of 0 was chosen almost always by the MI method. This was because when img2 was reset to img2 - a*img1, many extra unique pixel values are observed in the reset img2 vs. the original img2. This leads to higher entropy and therefore higher MI. As a result, whenever a>0, MI is inflated; thus, a=0 appears to minimize MI almost always. To combat this (and other artefacts of MI estimation), MI across all a values are replaced with their rolling mean with window size 11 (the original MI measurement and the 10 MI values for the 10 closest a values).